+++ example/example.dart
@@ -1,5 +1,4 @@
-import 'dart:io';
-
+import 'package:file/local.dart';
 import 'package:stash_disk/stash_disk.dart';
 
 class Task {
@@ -26,11 +25,14 @@ class Task {
 }
 
 void main() async {
+  // Local filesystem
+  final fs = const LocalFileSystem();
+
   // Temporary path
-  final path = Directory.systemTemp.path;
+  final path = fs.systemTempDirectory.path;
 
   // Creates cache with a disk based storage backend with the capacity of 10 entries
-  final cache = newDiskCache(path,
+  final cache = newDiskCache(const LocalFileSystem(), path,
       maxEntries: 10, fromEncodable: (json) => Task.fromJson(json));
 
   // Adds a task with key 'task1' to the cache