+++ lib/stash_disk.dart
@@ -1,6 +1,7 @@
 /// Provides a disk implementation of the Stash caching API for Dart
 library stash_disk;
 
+import 'package:file/file.dart';
 import 'package:stash/stash_api.dart';
 import 'package:stash_disk/src/disk/disk_store.dart';
 
@@ -8,6 +9,7 @@ export 'src/disk/disk_store.dart';
 
 /// Creates a new [Cache] backed by a [DiskStore]
 ///
+/// * [fs]: The [FileSystem]
 /// * [path]: The base storage location for this cache
 /// * [cacheName]: The name of the cache
 /// * [expiryPolicy]: The expiry policy to use, defaults to [EternalExpiryPolicy] if not provided
@@ -19,7 +21,7 @@ export 'src/disk/disk_store.dart';
 /// * [fromEncodable]: A custom function the converts to the object from a `Map<String, dynamic>` representation
 ///
 /// Returns a new [Cache] backed by a [DiskStore]
-Cache newDiskCache(String path,
+Cache newDiskCache(FileSystem fs, String path,
     {String cacheName,
     ExpiryPolicy expiryPolicy,
     KeySampler sampler,
@@ -29,7 +31,7 @@ Cache newDiskCache(String path,
     CacheCodec codec,
     dynamic Function(dynamic) fromEncodable}) {
   return Cache.newCache(
-      DiskStore(path, codec: codec, fromEncodable: fromEncodable),
+      DiskStore(fs, path, codec: codec, fromEncodable: fromEncodable),
       name: cacheName,
       expiryPolicy: expiryPolicy,
       sampler: sampler,